+++ FinanceApp.Shared/Dto/ForecastItem.cs
@@ -5,7 +5,7 @@ namespace FinanceApp.Shared.Dto
     public class ForecastItem
     {
         public DateTime DateReference { get; set;}
-        public decimal Amount { get; set;}
-        public decimal CumulatedAmount { get; set;}
+        public double Amount { get; set;}
+        public double CumulatedAmount { get; set;}
     }
 }
