+++ FinanceApp.Shared/Dto/Loan/CreateLoan.cs
@@ -7,8 +7,8 @@ namespace FinanceApp.Shared.Dto.Loan
         public DateTime InitialDate { get; set; }
         public int MonthsPayment { get; set; }
         public string Name { get; set; }
-        public decimal LoanValue { get; set; }
-        public decimal InterestRate { get; set; }
+        public double LoanValue { get; set; }
+        public double InterestRate { get; set; }
         public int Type { get; set; }
 
     }
