+++ FinanceApp.Shared/Dto/PrivateFixedInvestment/UpdatePrivateFixedIncome.cs
@@ -12,15 +12,15 @@ namespace FinanceApp.Shared.Dto.PrivateFixedInvestment
         [Required]
         public int Type { get; set; }
         [Required]
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         [Required]
         public int Index { get; set; }
         [Required]
         public bool PreFixedInvestment { get; set; }
         [Required]
-        public decimal IndexPercentage { get; set; }
+        public double IndexPercentage { get; set; }
         [Required]
-        public decimal AdditionalFixedInterest { get; set; }
+        public double AdditionalFixedInterest { get; set; }
         [Required]
         public DateTime InvestmentDate { get; set; }
         [Required]
