+++ FinanceApp.Shared/Dto/TreasuryBond/UpdateTreasuryBond.cs
@@ -6,10 +6,10 @@ namespace FinanceApp.Shared.Dto.TreasuryBond
     {
         public int Id { get; set; }
         public int Type { get; set; }
-        public decimal UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public DateTime ExpirationDate { get; set; }
         public DateTime InvestmentDate { get; set; }
         public int Operation { get; set; }
-        public decimal Quantity { get; set; }
+        public double Quantity { get; set; }
     }
 }
