+++ FinanceApp.Shared/Models/UserTables/Loan.cs
@@ -8,8 +8,8 @@ namespace FinanceApp.Shared.Models.UserTables
         public DateTime InitialDate { get; set; }
         public int MonthsPayment { get; set; }
         public string Name { get; set; }
-        public decimal LoanValue { get; set; }
-        public decimal InterestRate { get; set; }
+        public double LoanValue { get; set; }
+        public double InterestRate { get; set; }
         public EPaymentType Type { get; set; }
     }
 }
