+++ FinanceApp.Shared/Models/UserTables/PrivateFixedIncome.cs
@@ -11,15 +11,15 @@ namespace FinanceApp.Shared.Models.UserTables
         [Required]
         public ETypePrivateFixedIncome Type { get; set; }
         [Required]
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         [Required]
         public EIndex Index { get; set; }
         [Required]
         public bool PreFixedInvestment { get; set; }
         [Required]
-        public decimal IndexPercentage { get; set; }
+        public double IndexPercentage { get; set; }
         [Required]
-        public decimal AdditionalFixedInterest { get; set; }
+        public double AdditionalFixedInterest { get; set; }
         [Required]
         public DateTime InvestmentDate { get; set; }
         [Required]
